requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.4
dnspython==2.2.1
paho-mqtt==1.6.1
pycodestyle==2.10.0
pymongo==4.3.3
sqlparse==0.4.3
tomli==2.0.1
